Refactor and fix up styling
