+++ packages/apps/esm-login-app/src/change-location-link/change-location-link.scss
@@ -1,46 +1,13 @@
 @import '~@openmrs/esm-styleguide/src/vars';
 @import '../root.scss';
 
-.changeLocationLinkContainer {
+.panelItemContainer a {
   display: flex;
-  flex-direction: row;
-  width: 16rem;
-  justify-content: flex-start;
-  color: $field-01;
   align-items: center;
-  @extend .bodyLong01;
-  height: 3rem;
-  cursor: pointer;
-}
-
-.changeLocationLinkContainer svg {
-  margin: 0rem 0.75rem;
-  fill: $field-01;
-}
-
-.changeLocationLinkContainer div {
-  width: 100%;
-  display: flex;
   justify-content: space-between;
+}
+
+.panelItemContainer div {
+  display: flex;
   align-items: center;
-}
-
-.changeLocationLinkContainer button,
-.changeLocationLinkContainer button:hover {
-  color: $inverse-link;
-}
-
-.changeLocationLinkContainer button :hover {
-  @include brand-03(background-color);
-  color: $inverse-link;
-}
-
-.changeLocationLink {
-  color: $field-01;
-  width: fit-content;
-}
-
-.changeLocationLink :hover {
-  color: $field-01;
-  @include brand-01(background-color);
 }
