+++ packages/apps/esm-login-app/src/change-location-link/change-location-link.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ChangeLocationLink from './change-location-link.component';
+import ChangeLocationLink from './change-location-link.extension';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { navigate } from '@openmrs/esm-framework';
