+++ packages/apps/esm-login-app/src/logout/logout.extension.tsx
@@ -1,8 +1,9 @@
 import React, { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Button, Switcher, SwitcherDivider } from '@carbon/react';
+import { Button } from '@carbon/react';
 import { navigate } from '@openmrs/esm-framework';
 import styles from './logout.scss';
+import { SwitcherItem } from '@carbon/react';
 
 export interface LogoutProps {}
 
@@ -13,14 +14,11 @@ const Logout: React.FC<LogoutProps> = () => {
   }, []);
 
   return (
-    <>
-      <SwitcherDivider className={styles.divider} />
-      <Switcher aria-label="Switcher Container">
-        <Button className={styles.logout} onClick={logout} aria-labelledby="Logout" role="button">
-          {t('Logout', 'Logout')}
-        </Button>
-      </Switcher>
-    </>
+    <SwitcherItem>
+      <Button className={styles.logout} onClick={logout} aria-labelledby="Logout" kind="ghost">
+        {t('Logout', 'Logout')}
+      </Button>
+    </SwitcherItem>
   );
 };
