+++ packages/apps/esm-login-app/src/routes.json
@@ -27,11 +27,10 @@
     },
     {
       "name": "logout-button",
-      "slot": "user-panel-actions-slot",
+      "slot": "user-panel-bottom-slot",
       "component": "logoutButton",
       "online": true,
-      "offline": true,
-      "order": 999
+      "offline": true
     },
     {
       "name": "location-changer",
