+++ packages/apps/esm-offline-tools-app/src/offline-actions/offline-actions-mode-button.scss
@@ -1,38 +1,27 @@
 @import '~@openmrs/esm-styleguide/src/vars';
 @import '../root.scss';
 
-.offlineModeButtonContainer {
+.panelItemContainer a {
   display: flex;
-  flex-direction: row;
-  width: 16rem;
   justify-content: space-between;
-  color: $field-01;
   align-items: center;
-  @extend .bodyLong01;
-  height: 3rem;
-  cursor: pointer;
 
   :global(.cds--toggle-input__label .cds--toggle__switch) {
     margin-top: 0 !important;
   }
 
   :global(.cds--toggle) {
+    /* setting the width prevents the toggle from changing size when the text changes */
+    width: 76px;
     margin: 0 1rem;
   }
 
   :global(.cds--toggle__text) {
     color: $ui-02;
   }
+}
 
-  svg {
-    margin: 0rem 0.75rem;
-    fill: $field-01;
-  }
-
-  & > div {
-    display: flex;
-    flex-direction: row;
-    width: fit-content;
-    align-items: center;
-  }
+.panelItemContainer div {
+  display: flex;
+  align-items: center;
 }
