+++ packages/apps/esm-primary-navigation-app/src/components/change-language/change-language-link.extension.tsx
@@ -4,6 +4,7 @@ import { useTranslation } from 'react-i18next';
 import { showModal, useSession } from '@openmrs/esm-framework';
 import { Button } from '@carbon/react';
 import styles from './change-language-link.scss';
+import { SwitcherItem } from '@carbon/react';
 
 /** The user menu item that shows the current language and has a button to change the language */
 export function ChangeLanguageLink() {
@@ -17,15 +18,15 @@ export function ChangeLanguageLink() {
   const languageNames = new Intl.DisplayNames([session?.locale], { type: 'language' });
 
   return (
-    <div className={styles.changeLanguageLinkContainer}>
-      <Language size={20} />
+    <SwitcherItem className={styles.panelItemContainer}>
       <div>
-        {languageNames.of(session?.locale)}
-        <Button kind="ghost" onClick={launchChangeModal}>
-          {t('change', 'Change')}
-        </Button>
+        <Language size={20} />
+        <p>{languageNames.of(session?.locale)}</p>
       </div>
-    </div>
+      <Button kind="ghost" onClick={launchChangeModal}>
+        {t('change', 'Change')}
+      </Button>
+    </SwitcherItem>
   );
 }
