+++ packages/apps/esm-primary-navigation-app/src/components/navbar-header-panels/user-menu-panel.component.tsx
@@ -1,17 +1,19 @@
 import React from 'react';
-import type { LoggedInUser, Session } from '@openmrs/esm-framework';
-import { ExtensionSlot, useOnClickOutside } from '@openmrs/esm-framework';
 import type { HeaderPanelProps } from '@carbon/react';
-import { HeaderPanel } from '@carbon/react';
-import styles from '../../root.scss';
+import { HeaderPanel, Switcher, SwitcherDivider } from '@carbon/react';
+import { ExtensionSlot, useOnClickOutside } from '@openmrs/esm-framework';
+import styles from './user-menu-panel.scss';
 
 interface UserMenuPanelProps extends HeaderPanelProps {
   expanded: boolean;
-  onLogout(): void;
   hidePanel: () => void;
 }
 
-const UserMenuPanel: React.FC<UserMenuPanelProps> = ({ expanded, onLogout, hidePanel }) => {
+/**
+ * Extensions attaching to `user-panel-slot` or `user-panel-bottom-slot` should in
+ * general be wrapped in the `SwitcherItem` Carbon component.
+ */
+const UserMenuPanel: React.FC<UserMenuPanelProps> = ({ expanded, hidePanel }) => {
   const userMenuRef = useOnClickOutside<HTMLDivElement>(hidePanel, expanded);
 
   return (
@@ -22,13 +24,11 @@ const UserMenuPanel: React.FC<UserMenuPanelProps> = ({ expanded, onLogout, hideP
       aria-label="Location"
       aria-labelledby="Location Icon"
     >
-      <ExtensionSlot
-        name="user-panel-slot"
-        state={{
-          onLogout: onLogout,
-        }}
-      />
-      <ExtensionSlot name="user-panel-actions-slot" />
+      <Switcher className={styles.userPanelSwitcher} aria-label="Switcher Container">
+        <ExtensionSlot name="user-panel-slot" />
+        <SwitcherDivider className={styles.divider} />
+        <ExtensionSlot name="user-panel-bottom-slot" />
+      </Switcher>
     </HeaderPanel>
   );
 };
