+++ packages/apps/esm-primary-navigation-app/src/components/user-panel-switcher-item/user-panel-switcher.component.tsx
@@ -3,17 +3,16 @@ import { Switcher } from '@carbon/react';
 import { UserAvatarFilledAlt } from '@carbon/react/icons';
 import { useSession, type LoggedInUser } from '@openmrs/esm-framework';
 import styles from './user-panel-switcher.scss';
+import { SwitcherItem } from '@carbon/react';
 
 const UserPanelSwitcher: React.FC = () => {
   const session = useSession();
   const user = session?.user;
   return (
-    <div className={styles.switcherContainer}>
-      <Switcher aria-label="Switcher Container">
-        <UserAvatarFilledAlt size={20} />
-        <p>{user.person.display}</p>
-      </Switcher>
-    </div>
+    <SwitcherItem>
+      <UserAvatarFilledAlt size={20} />
+      <p>{user.person.display}</p>
+    </SwitcherItem>
   );
 };
