+++ packages/apps/esm-primary-navigation-app/src/components/user-panel-switcher-item/user-panel-switcher.scss
@@ -1,59 +1,7 @@
+@import '~@openmrs/esm-styleguide/src/vars';
 @import '../../root.scss';
 
-.switcherContainer {
+.panelItemContainer {
   display: flex;
-  flex-direction: column;
-}
-
-.switcherContainer ul {
-  display: flex;
-  flex-direction: row;
-  width: 20rem;
-  justify-content: flex-start;
-  color: $field-01;
   align-items: center;
-  @extend .productiveHeading01;
-  height: 3rem;
-  cursor: pointer;
-}
-
-.switcherContainer ul > svg {
-  margin: 0rem 0.75rem;
-  fill: $field-01;
-}
-
-.switcherContainer ul button {
-  @extend .productiveHeading01;
-  width: 100%;
-  background-color: var(--brand-02);
-}
-
-.switcherContainer ul button:hover {
-  color: $field-01;
-  background-color: var(--brand-01);
-}
-
-.switcherContainer ul div {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-start;
-  @extend .bodyLong01;
-}
-
-.switcherContainer ul div button {
-  color: $inverse-link;
-  margin-left: 0.5rem;
-  width: 5rem;
-}
-
-.switcherContainer ul div button:hover {
-  background-color: var(--brand-01);
-  color: $inverse-link;
-}
-
-.switcherContainer hr {
-  background-color: $ui-background;
-  width: 16rem;
-  margin: 0.5rem 0rem;
 }
